+++ rsvim_core/src/state/fsm.rs
@@ -44,6 +44,8 @@ pub mod terminal;
 pub mod visual;
 
 #[cfg(test)]
+mod insert_tests;
+#[cfg(test)]
 mod normal_tests;
 
 #[derive(Debug)]
